infrastructure/cdkv1/requirements.txt
aws-cdk.aws-ec2==1.156.1
aws_cdk.aws-fsx==1.156.1
aws_cdk.aws-batch==1.156.1
aws_cdk.aws-sagemaker==1.156.1
aws_cdk.aws-iam==1.156.1
aws_cdk.aws-kms==1.156.1
aws_cdk.aws-ecr==1.156.1
aws-cdk.aws-s3-deployment==1.156.1
aws-cdk.aws-s3-assets==1.156.1
python-dotenv==0.20.0
chalice[cdk]
-r "lokafold-app/requirements.txt"
# chalice[cdkv2] # extra to support chalice construct